Mark generated CREATE events as out of band.
Deliver such events to each handler that receives CREATE and WRITE
events.  This makes sure all handler have a chance to notice creation
of a file within the time that expires between  creation of a directory
and creation of a watchpoint for it.  Introduce new marker (flag) for
such out of band data.

Make sure out of band CREATE notifications are sent to the deepest
created watchpoint, in case the recursion depth is limited.

* src/direvent.h (GENEV_OOB): New macro for out of band events.
* src/event.c (genev_transtab): Add an entry for GENEV_OOB.
* src/watcher.c (deliver_ev_create): Deliver out of band CREATE
event.  Receipients are all handlers that requested CREATE or WRITE
events.
(check_new_watcher): Fix misleading comment.
(watch_subdirs):  Make sure out of band CREATE notifications are sent
if the watchpoint has zero depth (last directory in a depth-limited
recusive watchpoint chain).
* tests/createrec2.at: Remove sleeps between creation of a directory
and creation of a file in it.